# lib/api-spec/openapi_half.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PMFBY Crop Insurance AI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /predictions:
    get:
      operationId: listPredictions
      tags: [predictions]
      summary: List AI model predictions
      parameters:
        - name: claimId
          in: query
          schema: { type: integer }
        - name: cropType
          in: query
          schema: { type: string }
        - name: flaggedOnly
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prediction"
    post:
      operationId: createPrediction
      tags: [predictions]
      summary: Run AI prediction on crop image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictionInput"
      responses:
        "201":
          description: Prediction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prediction"
  /predictions/{id}:
    get:
      operationId: getPrediction
      tags: [predictions]
      summary: Get prediction details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Prediction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prediction"
        "404":
          description: Not found
...
  /weather-events:
    get:
      operationId: listWeatherEvents
      tags: [weather]
      summary: List weather events for cross-validation
      parameters:
        - name: district
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: eventType
          in: query
          schema: { type: string }
        - name: fromDate
          in: query
          schema: { type: string, format: "date" }
        - name: toDate
          in: query
          schema: { type: string, format: "date" }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of weather events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeatherEvent"
    post:
      operationId: createWeatherEvent
      tags: [weather]
      summary: Log a weather event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeatherEventInput"
      responses:
        "201":
          description: Weather event logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherEvent"

  # ─── Dashboard ────────────────────────────────────────────────────────────
...